+++ src/com/driveventures/service/UsuarioService.java
@@ -9,7 +9,7 @@ import DBCUtils.MailException;
 
 public interface UsuarioService {
 
-	public Usuario findById(int id) throws DataException;
+	public Usuario findById(Long id) throws DataException;
 	
 	public Usuario login(String email, String password) throws DataException;
